1: refactor: combin old 'localClinet' and 'jwtClient' into 'AuthClinet' 2: integration tests for authclient.
